net/qed: allow old cards not supporting "num_images" to work
Commit 43645ce03e0063d7c4a5001215ca815188778881 added support for
populating flash image attributes, notably "num_images". However, some
cards were not able to return this information. In such cases, the
driver would return EINVAL, causing the driver to exit.

We added a check to return EOPNOTSUPP when the card is not able to
return these information, allowing the driver continue instead of
returning an error.

Co-developed-by: Florian Forestier <florian@forestier.re>
Signed-off-by: Florian Forestier <florian@forestier.re>
Signed-off-by: Louis Leseur <louis.leseur@gmail.com>